+++ Proyecto Java/src/ejecucion.java
@@ -22,8 +22,8 @@ public class ejecucion {
 		str1=separar.espacios(str[0]);
 		str2=separar.espacios(str[2]);
 
-		if(str1[1].equals("alumno")&&(avisos.fechaingreso(str2[1],str2[2])&avisos.notamedia(Integer.parseInt(str2[3].trim()))&avisos.imp())){
-			mapa.put(getid(mapa),new alumno(str[1], getid(mapa),str2[1],str1[1],str2[2],Integer.parseInt(str2[3].trim()),"", ""));
+		if(str1[1].equals("alumno")&&(avisos.fechaingreso(str2[1],str2[2])&avisos.notamedia(Double.parseDouble(str2[3].trim()))&avisos.imp())){
+			mapa.put(getid(mapa),new alumno(str[1], getid(mapa),str2[1],str1[1],str2[2],Double.parseDouble(str2[3].trim()),"", ""));
 		}
 		if(str1[1].equals("profesor")){
 			str3=separar.espacios(str[4]);
